chore: file & fn renames for clarity
