+++ crates/common/src/lib.rs
@@ -3,7 +3,7 @@ pub mod deposit;
 pub mod error;
 pub mod forked_evm_cache;
 pub mod frontmatter;
-pub mod lsp_services;
+pub mod dotrain_add_order_lsp;
 pub mod rainlang;
 pub mod remove_order;
 pub mod subgraph;
