+++ tauri-app/src-tauri/src/commands/dotrain_add_order_lsp.rs
@@ -1,17 +1,17 @@
 use rain_orderbook_common::{
     dotrain::types::ast::Problem,
     dotrain_lsp::lsp_types::{CompletionItem, Hover, Position, TextDocumentItem},
-    lsp_services::{DotrainAddOrderLsp},
+    dotrain_add_order_lsp::DotrainAddOrderLsp,
 };
 use crate::error::CommandResult;
 
 #[tauri::command]
-pub fn provide_hover(text_document: TextDocumentItem, position: Position) -> Option<Hover> {
+pub fn call_lsp_hover(text_document: TextDocumentItem, position: Position) -> Option<Hover> {
     DotrainAddOrderLsp::new(text_document).hover(position)
 }
 
 #[tauri::command]
-pub fn provide_completion(
+pub fn call_lsp_completion(
     text_document: TextDocumentItem,
     position: Position,
 ) -> Option<Vec<CompletionItem>> {
@@ -19,7 +19,7 @@ pub fn provide_completion(
 }
 
 #[tauri::command]
-pub async fn provide_problems(
+pub async fn call_lsp_problems(
     text_document: TextDocumentItem,
     rpc_url: &str,
     block_number: Option<u64>,
