+++ tauri-app/src-tauri/src/commands/mod.rs
@@ -1,6 +1,6 @@
 pub mod chain;
 pub mod dotrain;
-pub mod lsp_services;
+pub mod dotrain_add_order_lsp;
 pub mod order;
 pub mod order_clear;
 pub mod vault;
